+++ public/style.css
@@ -57,11 +57,11 @@ body ::-webkit-scrollbar {
 
 @media only screen and (max-width: 768px) {
   body .container {
-    padding-top: 0;
+    padding-top: 1.2em;
     height: 100vh;
   }
   .friends-list {
-    height: 76.5vh;
+    height: 77.5vh;
     align-items: baseline;
     align-content: baseline;
   }
@@ -70,11 +70,11 @@ body ::-webkit-scrollbar {
   }
   .dashboard-card {
     padding: 0;
-    /* height: 82vh; */
+    height: 82vh;
   }
-  /* .mobile {
-    margin-top: 12vh;
-  } */
+  .mobile {
+    height: 90vh;
+  }
   .ui.card {
     margin: 0em 0;
     margin-top: 12vh;
